+++ src/SharpX/Extensions/StringExtensions.cs
@@ -29,7 +29,7 @@ namespace SharpX
         static string[] _mangleChars =
             {"!", "\"", "£", "$", "%", "&", "/", "(", ")", "=", "?", "^", "[", "]", "*", "@", "°",
              "#", "§", ",", ";", ".", ":", "-", "_"};
-        static Regex _stripTagsRegEx = new Regex(@"<[^>]*>", RegexOptions.Compiled | RegexOptions.Multiline);
+        static Regex _stripTagRegEx = new Regex(@"<[^>]*>", RegexOptions.Compiled | RegexOptions.Multiline);
 
         /// <summary>Determines if a string is composed only by letter characters.</summary>
         public static bool IsAlpha(this string value)
@@ -231,11 +231,11 @@ namespace SharpX
         }
 
         /// <summary>Removes tags from a string.</summary>
-        public static string StripTags(this string value)
+        public static string StripTag(this string value)
         {
             if (value == null) throw new ArgumentNullException(nameof(value));
 
-            return _stripTagsRegEx.Replace(value, string.Empty);
+            return _stripTagRegEx.Replace(value, string.Empty);
         }
 
         /// <summary>Removes words of given length.</summary>